Added fullscreen support on all OS
