+++ skiko/src/jvmMain/cpp/linux/drawlayer.cc
@@ -93,6 +93,10 @@ private:
 set<LayerHandler *> *layerStorage = NULL;
 LayerHandler *findByObject(JNIEnv *env, jobject object)
 {
+    if (layerStorage == NULL)
+    {
+        return NULL;
+    }
     for (auto &layer : *layerStorage)
     {
         if (env->IsSameObject(object, layer->canvasGlobalRef) == JNI_TRUE)
@@ -106,7 +110,7 @@ LayerHandler *findByObject(JNIEnv *env, jobject object)
 
 extern "C"
 {
-    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_HardwareLayer_updateLayer(JNIEnv *env, jobject canvas)
+    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_layer_HardwareLayer_updateLayer(JNIEnv *env, jobject canvas)
     {
         if (layerStorage != NULL)
         {
@@ -169,7 +173,7 @@ extern "C"
         awt.FreeDrawingSurface(ds);
     }
 
-    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_HardwareLayer_redrawLayer(JNIEnv *env, jobject canvas)
+    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_layer_HardwareLayer_redrawLayer(JNIEnv *env, jobject canvas)
     {
         LayerHandler *layer = findByObject(env, canvas);
         if (layer != NULL)
@@ -178,7 +182,7 @@ extern "C"
         }
     }
 
-    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_HardwareLayer_disposeLayer(JNIEnv *env, jobject canvas)
+    JNIEXPORT void JNICALL Java_org_jetbrains_skiko_layer_HardwareLayer_disposeLayer(JNIEnv *env, jobject canvas)
     {
         LayerHandler *layer = findByObject(env, canvas);
         if (layer != NULL)
@@ -189,7 +193,7 @@ extern "C"
         }
     }
 
-    JNIEXPORT jlong JNICALL Java_org_jetbrains_skiko_HardwareLayer_getWindowHandle(JNIEnv *env, jobject canvas)
+    JNIEXPORT jlong JNICALL Java_org_jetbrains_skiko_layer_HardwareLayer_getWindowHandle(JNIEnv *env, jobject canvas)
     {
         JAWT awt;
         JAWT_DrawingSurface *ds = NULL;
