+++ skiko/src/jvmMain/kotlin/org/jetbrains/skiko/layer/Drawable.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skiko
+package org.jetbrains.skiko.layer
 
 interface Drawable {
     fun redrawLayer()
