+++ skiko/src/jvmMain/kotlin/org/jetbrains/skiko/properties/PlatformProperties.kt
@@ -1,0 +1,8 @@
+package org.jetbrains.skiko.properties
+
+import java.awt.Component
+
+internal interface PlatformProperties {
+    fun isFullscreen(component: Component): Boolean
+    fun makeFullscreen(component: Component, value: Boolean)
+}
