+++ skiko/src/jvmMain/kotlin/org/jetbrains/skiko/render/GraphicsApi.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skiko
+package org.jetbrains.skiko.render
 
 enum class GraphicsApi {
     UNKNOWN, OPENGL, VULKAN, METAL
